Move teams helper functions into lib/utils.
